+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/PresentationFactory.kt
@@ -7,7 +7,6 @@ import at.asitplus.dif.FormatHolder
 import at.asitplus.dif.PresentationDefinition
 import at.asitplus.jsonpath.JsonPath
 import at.asitplus.openid.*
-import at.asitplus.openid.OpenIdConstants.Errors
 import at.asitplus.openid.OpenIdConstants.VP_TOKEN
 import at.asitplus.signum.indispensable.CryptoPublicKey
 import at.asitplus.signum.indispensable.cosef.CoseSigned
@@ -26,6 +25,7 @@ import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.iso.*
 import at.asitplus.wallet.lib.jws.JwsService
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception
+import at.asitplus.wallet.lib.oidvci.OAuth2Exception.*
 import io.github.aakira.napier.Napier
 import io.matthewnelson.encoding.base16.Base16
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
@@ -73,7 +73,7 @@ internal class PresentationFactory(
             credentialPresentation = credentialPresentation,
         ).getOrElse {
             Napier.w("Could not create presentation", it)
-            throw OAuth2Exception(Errors.USER_CANCELLED, it)
+            throw AccessDenied("Could not create presentation", it)
         }.also { presentation ->
             clientMetadata?.vpFormats?.let {
                 when (presentation) {
@@ -177,7 +177,7 @@ internal class PresentationFactory(
         }
         val nonce = request.parameters.nonce ?: run {
             Napier.w("nonce is null in ${request.parameters}")
-            throw OAuth2Exception(Errors.INVALID_REQUEST, "nonce is null")
+            throw InvalidRequest("nonce is null")
         }
         val now = clock.now()
         // we'll assume jwk-thumbprint
@@ -200,7 +200,7 @@ internal class PresentationFactory(
             addX5c = false
         ).getOrElse {
             Napier.w("Could not sign id_token", it)
-            throw OAuth2Exception(Errors.USER_CANCELLED, it)
+            throw AccessDenied("Could not sign id_token", it)
         }
     }
 
@@ -208,7 +208,7 @@ internal class PresentationFactory(
     private fun RequestParameters.verifyResponseType() {
         if (responseType == null || !responseType!!.contains(VP_TOKEN)) {
             Napier.w("vp_token not requested in response_type='$responseType'")
-            throw OAuth2Exception(Errors.INVALID_REQUEST, "response_type invalid")
+            throw InvalidRequest("response_type invalid")
         }
     }
 
@@ -221,14 +221,14 @@ internal class PresentationFactory(
         val validator = PresentationSubmissionValidator.createInstance(this).getOrThrow()
         if (!validator.isValidSubmission(credentialSubmissions.keys)) {
             Napier.w("submission requirements are not satisfied")
-            throw OAuth2Exception(Errors.USER_CANCELLED, "submission requirements not satisfied")
+            throw UserCancelled("submission requirements not satisfied")
         }
 
         // making sure, that all the submissions actually match the corresponding input descriptor requirements
         credentialSubmissions.forEach { submission ->
             val inputDescriptor = this.inputDescriptors.firstOrNull { it.id == submission.key } ?: run {
                 Napier.w("Invalid input descriptor id: ${submission.key}")
-                throw OAuth2Exception(Errors.USER_CANCELLED, "invalid input_descriptor_id")
+                throw UserCancelled("invalid input_descriptor_id")
             }
 
             val constraintFieldMatches = holder.evaluateInputDescriptorAgainstCredential(
@@ -251,7 +251,7 @@ internal class PresentationFactory(
                 disclosedAttributes.firstOrNull { allowedPaths.contains(it) } ?: run {
                     val keyId = constraintField.key.id?.let { " Missing field: $it" }
                     Napier.w("Input descriptor constraints not satisfied: ${inputDescriptor.id}.$keyId")
-                    throw OAuth2Exception(Errors.USER_CANCELLED, "constraints not satisfied")
+                    throw UserCancelled("constraints not satisfied")
                 }
             }
             // TODO: maybe we also want to validate, whether there are any redundant disclosed attributes?
@@ -266,7 +266,7 @@ internal class PresentationFactory(
         presentationSubmission.descriptorMap?.mapIndexed { _, descriptor ->
             if (supportedFormats.isMissingFormatSupport(descriptor.format)) {
                 Napier.w("Incompatible JWT algorithms for claim format ${descriptor.format}: $supportedFormats")
-                throw OAuth2Exception(Errors.REGISTRATION_VALUE_NOT_SUPPORTED, "incompatible algorithms")
+                throw RegistrationValueNotSupported("incompatible algorithms")
             }
         }
 
@@ -280,7 +280,7 @@ internal class PresentationFactory(
             }
             if (supportedFormats.isMissingFormatSupport(format)) {
                 Napier.w("Incompatible JWT algorithms for claim format $format: $supportedFormats")
-                throw OAuth2Exception(Errors.REGISTRATION_VALUE_NOT_SUPPORTED, "incompatible algorithms")
+                throw RegistrationValueNotSupported("incompatible algorithms")
             }
         }